Fix Xjoin test on Python >=3.7
Python 3.7 changed re.escape to escape only regular expression control
characters. That means that non-ASCII characters and some special
ASCII characters like ! or / are no longer escaped. In one test, the
escaped string was passed as a literal, not itself using the re.escape
function. Fixed by reusing re.escape in the test.

This is still not 100% correct, because we are using Python re.escape
to escape a regular expression that is passed to Crossjoin. The
Inventory does not know what regex engine does Crossjoin use and it
should not matter at all. Crossjoin should describe its regex engine
including what characters need to be escaped. The current behavior
incorrectly assumes that whatever engine Crossjoin uses, it is
compatible with the Python engine. This is a potential problem if
Crossjoin uses or starts using an engine that uses control characters
not considered as such by re.escape.