Check-In endpoint refactoring and cleanup (#744)
* Remove unused fixture; simplify api_post

* Refactor canonical facts in API spec

* Allow canonical fact fields to be nullable

* Return 400 status if no canonical facts are provided

* Standardize on flask.abort for errors

* Separate CanonicalFactsIn and CanonicalFactsOut

* DRY Canonical Facts parts

Extracted canonical facts parts into their own components that are
reused in CanonicalFactsOut and CanonicalFactsIn.

* DRY CanonicalFactsIn and CanonicalFactsOut

CanonicalFactsIn is a subset of CanonicalFactsOut. Use allOf to
reference the whole CanonicalFactsOut and add the non-nullable
constraints all at once.

* Make at least one Canonical Fact mandatory

Used a combination of oneOf and required to make at least one
Canonical Fact required purely by the Schema.

* Remove Bad Request from defined /checkin responses

* Allow multiple canonical facts in request; parametrize test

* Parametrize no Canonical Facts test

Modified the test that verifies that the Check-in request fails if no
Canonical Facts are provided. Parametrized it so it checks both
variants – with and without checkin frequency.

* Add tests for checkin frequency

Added tests that verify that the checkin frequency value is received,
validated and passed to the operation function.

* Remove checkin frequency default

Connexion uses default values only for parameters, not for request
body. Like that, the value there is misleading as it does not control
Connexion and there is no other code that would fill in the default
values.

* Remove checkin frequency

The checkin frequency property is not used at all. Having it in the
specification is only misleading. Removed completely together with its
tests.

* Revert "Remove checkin frequency"

This reverts commit 9846e86133af207ab0101528ecb6bc25de9625fd.

Co-authored-by: Glutexo <stomsa@redhat.com>