fix: add yum update to fix security vulnerabilities at build
This is a method I use with a few of the apps in the platform to keep
the packages on our images up to date. I noticed some alerts in the quay
vulnerability channel in slack. You can also update, or remove packages
you don't use, but this is handy at keeping images updated at build.

Signed-off-by: Stephen Adams <tsadams@gmail.com>