Add a /hosts/stats endpoint with counts for the Dashboard
At the moment, the Dashboard makes these three requests of the Inventory API:

- https://console.redhat.com/api/inventory/v1/hosts?registered_with=puptoo&per_page=1
- https://console.redhat.com/api/inventory/v1/hosts?staleness=stale&registered_with=puptoo&per_page=1
- https://console.redhat.com/api/inventory/v1/hosts?staleness=stale_warning&registered_with=puptoo&per_page=1

These in turn make three separate requests to XJoin to get that data.
We'd like to reduce that to one request to:

- https://console.redhat.com/api/inventory/v1/hosts/stats

This would produce data in a JSON object with keys:

  - total: {total of puptoo-registered hosts}
  - stale: {number of stale puptoo-registered hosts}
  - warning: {number of stale-warning puptoo-registered hosts}

The code so far does this in three SQL queries, but hopefully it will be
possible to reduce this to one query.

Signed-off-by: Paul Wayper <paulway@redhat.com>