refactor[ESSNTL 3704]: group is only added to db if there's no problem with hosts
Group creation and host -> group assoc creation now happens within the same session,
meaning that all the changes are only going to be committed to the db if nothing goes wrong in the entire process.

Before, if a problem happened when creation a host -> group association, the group would still be created regardless.

Related: ESSNTL-3704, ESSNTL-4390