[RHINENG-9943] Allow to disable xjoin by configuration
This PR allows to disable `xjoin` by using a new environment variable.
Using the new variable `BYPASS_XJOIN=true` and `BYPASS_UNLEASH=true`,
forces the use of DB queries to get hosts. Using BYPASS_UNLEASH=true by
itself makes host-inventory to fallback to pre-unleash settings that
required xjoin, which is not easy to setup in a dev environment.

With this fix in place, unleash container is not even needed. If this
change is acceptable and is merged, the README will be updated.

Related-Bug: RHINENG-9943
Co-Authored-By: Muhammad Arif <aarif@redhat.com>
Signed-off-by: Gaël Chamoulaud <gchamoul@redhat.com>