Fix(RHINENG-21108): add local development server to API spec and openapi.json
When using the `dev.yml` docker compose file, we have access to a
swagger UI instance for API testing purpose. The swagger UI was missing
the path to our API (`api/inventory/v1`) in the url used to get data,
which does make the swagger UI unusable.

This commits fixes that by adding a servers section to the swagger spec.

To test it:
```
$ docker-compose -f dev.yml up

open a browser and go to http://localhost:8080/api/inventory/v1/ui/#/
Authorize yourself with the x-rh-identity and try to play with one endpoint.
```